Renames users-grid-test to users-grid
