+++ src/blocks.js
@@ -16,5 +16,4 @@ import './blocks/hero';
 import './blocks/related-posts';
 import './blocks/two-column-block';
 import './blocks/recent-posts';
-// import './blocks/users-grid';
-import './blocks/users-grid-test';
+import './blocks/users-grid';
